Created: Add Gast::Base. Move from function of initialize to Gast::Base.
