+++ lib/gast.rb
@@ -6,6 +6,7 @@ require 'sprockets-helpers'
 require 'gast/memo'
 require 'gast/app'
 require 'gast/repository'
+require 'gast/base'
 require 'haml'
 require 'sass'
 require 'git'
